+++ R/models.R
@@ -59,16 +59,62 @@
 #'
 NULL
 
+#' Indicate model type to lime
+#'
+#' `lime` requires knowledge about the type of model it is dealing with, more
+#' specifically whether the model is a regressor or a classifier. If the model
+#' class has a [model_type()] method defined lime can figure it out on its own
+#' but if not, you can wrap your model in either of these functions to indicate
+#' what type of model lime is dealing with. This can also be used to overwrite
+#' the output from [model_type()] if the implementation uses some heuristic that
+#' doesn't work for your particular model (e.g. keras models types are found by
+#' checking if the activation in the last layer is linear or not - this is
+#' rather crude). In addition `as_classifier` can be used to overwrite the
+#' returned class labels - this is handy if the model does not store the labels
+#' (again, keras springs to mind).
+#'
+#' @param x The model object
+#' @param labels An optional character vector giving labels for each class
+#'
+#' @return A model augmented with information about the model type and
+#' (potentially) the class labels.
+#'
+#' @export
+as_classifier <- function(x, labels = NULL) {
+  class(x) <- c('lime_classifier', class(x))
+  attr(x, 'lime_labels') <- labels
+  x
+}
+#' @rdname as_classifier
+#' @export
+as_regressor <- function(x) {
+  class(x) <- 'lime_regressor'
+  x
+}
+set_labels <- function(res, model) {
+  labels <- attr(model, 'lime_labels')
+  if (model_type(model) == 'classification' && !is.null(labels)) {
+    if (length(labels) != ncol(res)) {
+      warning('Ignoring provided class labels as length differs from model output')
+    } else {
+      names(res) <- labels
+    }
+  }
+  res
+}
+
 #' @rdname model_support
 #' @export
 predict_model <- function(x, newdata, type, ...) {
   UseMethod('predict_model')
 }
+#' @export
 predict_model.default <- function(x, newdata, type, ...) {
   p <- predict(x, newdata = newdata, type = type, ...)
   if (type == 'raw') p <- data.frame(Response = p, stringsAsFactors = FALSE)
   as.data.frame(p)
 }
+#' @export
 predict_model.WrappedModel <- function(x, newdata, type, ...) {
   if (!requireNamespace('mlr', quietly = TRUE)) {
     stop('mlr must be available when working with WrappedModel models')
@@ -81,6 +127,7 @@ predict_model.WrappedModel <- function(x, newdata, type, ...) {
     stop('Type must be either "raw" or "prob"', call. = FALSE)
   )
 }
+#' @export
 predict_model.xgb.Booster <- function(x, newdata, type, ...) {
   if (!requireNamespace('xgboost', quietly = TRUE)) {
     stop('The xgboost package is required for predicting xgboost models')
@@ -101,6 +148,7 @@ predict_model.xgb.Booster <- function(x, newdata, type, ...) {
   }
   p
 }
+#' @export
 predict_model.lda <- function(x, newdata, type, ...) {
   res <- predict(x, newdata = newdata, ...)
   switch(
@@ -109,33 +157,57 @@ predict_model.lda <- function(x, newdata, type, ...) {
     prob = as.data.frame(res$posterior, check.names = FALSE)
   )
 }
+#' @export
+predict_model.keras.engine.training.Model <- function(x, newdata, type, ...) {
+  if (!requireNamespace('keras', quietly = TRUE)) {
+    stop('The keras package is required for predicting keras models')
+  }
+  res <- predict(x, newdata)
+  if (type == 'raw') {
+    data.frame(Response = res[, 1])
+  } else {
+    if (ncol(res) == 1) {
+      res <- cbind(1 - res, res)
+    }
+    colnames(res) <- as.character(seq_len(ncol(res)))
+    as.data.frame(res, check.names = FALSE)
+  }
+}
+#' @export
 predict_model.H2OModel <- function(x, newdata, type, ...){
-    if (!requireNamespace('h2o', quietly = TRUE)) {
-        stop('The h2o package is required for predicting h2o models')
-    }
-    pred <- h2o::h2o.predict(x, h2o::as.h2o(newdata))
-    h2o_model_class <- class(x)[[1]]
-    if (h2o_model_class %in% c("H2OBinomialModel", "H2OMultinomialModel")) {
-        return(as.data.frame(pred[,-1]))
-    } else if (h2o_model_class == "H2ORegressionModel") {
-        ret <- as.data.frame(pred[,1])
-        names(ret) <- "Response"
-        return(ret)
-    } else {
-        stop('This h2o model is not currently supported.')
-    }
+  if (!requireNamespace('h2o', quietly = TRUE)) {
+      stop('The h2o package is required for predicting h2o models')
+  }
+  pred <- h2o::h2o.predict(x, h2o::as.h2o(newdata))
+  h2o_model_class <- class(x)[[1]]
+  if (h2o_model_class %in% c("H2OBinomialModel", "H2OMultinomialModel")) {
+      return(as.data.frame(pred[,-1]))
+  } else if (h2o_model_class == "H2ORegressionModel") {
+      ret <- as.data.frame(pred[,1])
+      names(ret) <- "Response"
+      return(ret)
+  } else {
+      stop('This h2o model is not currently supported.')
+  }
 }
 #' @rdname model_support
 #' @export
 model_type <- function(x, ...) {
   UseMethod('model_type')
 }
+#' @export
 model_type.default <- function(x, ...) {
   stop('The class of model must have a model_type method. See ?model_type to get an overview of models supported out of the box', call. = FALSE)
 }
+#' @export
+model_type.lime_classifier <- function(x, ...) 'classification'
+#' @export
+model_type.lime_regressor <- function(x, ...) 'regression'
+#' @export
 model_type.train <- function(x, ...) {
   tolower(x$modelType)
 }
+#' @export
 model_type.WrappedModel <- function(x, ...) {
   switch(
     x$learner$type,
@@ -146,6 +218,7 @@ model_type.WrappedModel <- function(x, ...) {
     multilabel = 'multilabel'
   )
 }
+#' @export
 model_type.xgb.Booster <- function(x, ...) {
   obj <- x$params$objective
   type <- strsplit(obj, ':')[[1]][1]
@@ -157,14 +230,27 @@ model_type.xgb.Booster <- function(x, ...) {
     stop('Unsupported model type', call. = FALSE)
   )
 }
+#' @export
 model_type.lda <- function(x, ...) 'classification'
+#' @export
+model_type.keras.engine.training.Model <- function(x, ...) {
+  if (!requireNamespace('keras', quietly = TRUE)) {
+    stop('The keras package is required for predicting keras models')
+  }
+  if (keras::get_layer(x, index = -1)$activation$func_name == 'linear') {
+    'regression'
+  } else {
+    'classification'
+  }
+}
+#' @export
 model_type.H2OModel <- function(x, ...) {
-    h2o_model_class <- class(x)[[1]]
-    if (h2o_model_class %in% c("H2OBinomialModel", "H2OMultinomialModel")) {
-        return('classification')
-    } else if (h2o_model_class == "H2ORegressionModel") {
-        return('regression')
-    } else {
-        stop('This h2o model is not currently supported.')
-    }
+  h2o_model_class <- class(x)[[1]]
+  if (h2o_model_class %in% c("H2OBinomialModel", "H2OMultinomialModel")) {
+      return('classification')
+  } else if (h2o_model_class == "H2ORegressionModel") {
+      return('regression')
+  } else {
+      stop('This h2o model is not currently supported.')
+  }
 }